Refactor: merge ConsistentHashing and RoundRobin to the same group
