+++ adapters/outboundgroup/loadbalance.go
@@ -17,8 +17,17 @@ import (
 type LoadBalance struct {
 	*outbound.Base
 	single    *singledo.Single
-	maxRetry  int
 	providers []provider.ProxyProvider
+}
+
+type ConsistentHashing struct {
+	*LoadBalance
+	maxRetry int
+}
+
+type RoundRobin struct {
+	*LoadBalance
+	index int
 }
 
 func getKey(metadata *C.Metadata) string {
@@ -81,21 +90,6 @@ func (lb *LoadBalance) SupportUDP() bool {
 	return true
 }
 
-func (lb *LoadBalance) Unwrap(metadata *C.Metadata) C.Proxy {
-	key := uint64(murmur3.Sum32([]byte(getKey(metadata))))
-	proxies := lb.proxies()
-	buckets := int32(len(proxies))
-	for i := 0; i < lb.maxRetry; i, key = i+1, key+1 {
-		idx := jumpHash(key, buckets)
-		proxy := proxies[idx]
-		if proxy.Alive() {
-			return proxy
-		}
-	}
-
-	return proxies[0]
-}
-
 func (lb *LoadBalance) proxies() []C.Proxy {
 	elm, _, _ := lb.single.Do(func() (interface{}, error) {
 		return getProvidersProxies(lb.providers), nil
@@ -115,11 +109,52 @@ func (lb *LoadBalance) MarshalJSON() ([]byte, error) {
 	})
 }
 
-func NewLoadBalance(name string, providers []provider.ProxyProvider) *LoadBalance {
-	return &LoadBalance{
-		Base:      outbound.NewBase(name, "", C.LoadBalance, false),
-		single:    singledo.NewSingle(defaultGetProxiesDuration),
-		maxRetry:  3,
-		providers: providers,
+func (ch *ConsistentHashing) Unwrap(metadata *C.Metadata) C.Proxy {
+	key := uint64(murmur3.Sum32([]byte(getKey(metadata))))
+	proxies := ch.proxies()
+	buckets := int32(len(proxies))
+	for i := 0; i < ch.maxRetry; i, key = i+1, key+1 {
+		idx := jumpHash(key, buckets)
+		proxy := proxies[idx]
+		if proxy.Alive() {
+			return proxy
+		}
+	}
+
+	return proxies[0]
+}
+
+func (rr *RoundRobin) Unwrap(metadata *C.Metadata) C.Proxy {
+	proxies := rr.proxies()
+	for i := 0; i < len(proxies); i++ {
+		rr.index = (rr.index + 1) % len(proxies)
+		proxy := proxies[rr.index]
+		if proxy.Alive() {
+			return proxy
+		}
+	}
+
+	return proxies[0]
+}
+
+func NewConsistentHashing(name string, providers []provider.ProxyProvider) *ConsistentHashing {
+	return &ConsistentHashing{
+		LoadBalance: &LoadBalance{
+			Base:      outbound.NewBase(name, "", C.LoadBalance, false),
+			single:    singledo.NewSingle(defaultGetProxiesDuration),
+			providers: providers,
+		},
+		maxRetry:    3,
+	}
+}
+
+func NewRoundRobin(name string, providers []provider.ProxyProvider) *RoundRobin {
+	return &RoundRobin{
+		LoadBalance: &LoadBalance{
+			Base:      outbound.NewBase(name, "", C.LoadBalance, false),
+			single:    singledo.NewSingle(defaultGetProxiesDuration),
+			providers: providers,
+		},
+		index:       0,
 	}
 }
